Create README.md
This is my First year second semester project. This project is mainly based on classes and includes the concepts of OOP and Pointers. It visualizes how social network apps work and this is only console based. The text files includes all the data of the users and the pages. In the command.txt it includes all the commands which can be performed by the app. I've made a run function which has some commands hardcoded in it so you can get the idea how they work and you can only give commands manually by uncommenting the while loop and commenting the run function in the main function.